Address PR review: split tests, add visual tests, fix mouse drag, remove legend styling
Agent-Logs-Url: https://github.com/gui-cs/Terminal.Gui/sessions/9d0abd93-a506-4979-85bb-22af119cf919
