+++ Tests/UnitTestsParallelizable/Views/LinearRange/LinearRangeDefaultKeyBindingsTests.cs
